fix(mimetype): fix missing charset on certain non-text mime types
There was an issue where UTF-8 files were interpreted in other charset because their content-type not specified the UTF-8 charset and the HTML had other encoding.

In our case a JS file was uploaded to S3 with "content-type: application/javascript" without charset=UTF-8 and it caused issues on www env because it uses iso-8859-1.

resolve WMA-456

Co-authored-by: bmori <balint.mori@emarsys.com>